refactor: :recycle: move logic inside cheatsheet
