+++ components/CheatSheet/CheatSheet.tsx
@@ -1,20 +1,34 @@
 import { useTranslation } from "next-i18next";
+import { useEffect, useState } from "react";
+import { useActiveItems } from "../../hooks/useActiveItems";
 import * as S from "./CheatSheet.atoms";
 import { Item, ItemProps } from "./Item/Item";
 
-interface CheatSheetProps {
-  items: ItemProps[];
-  active: Record<string, boolean>;
-  onClose: () => void;
-}
-
-export const CheatSheet = ({ items, active, onClose }: CheatSheetProps) => {
+export const CheatSheet = () => {
+  const { ITEMS: items, active } = useActiveItems();
   const { t } = useTranslation(["common"], { keyPrefix: "cheatSheet" });
+  const [showCheatSheet, setShowCheatSheet] = useState(false);
+
+  useEffect(() => {
+    const infoToggle = (e: KeyboardEvent) => {
+      if (e.ctrlKey && e.key === "i") {
+        setShowCheatSheet((prev) => !prev);
+      }
+    };
+
+    document.addEventListener("keydown", infoToggle);
+
+    return () => {
+      document.removeEventListener("keydown", infoToggle);
+    };
+  }, []);
+
+  if (!showCheatSheet) return null;
 
   return (
     <S.Container>
       <S.Title>{t("title")}</S.Title>
-      <S.Close onClick={onClose}>
+      <S.Close onClick={() => setShowCheatSheet(false)}>
         <S.Cross src="/cross.png" />
       </S.Close>
       <S.Grid>
